+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `isArchived` to Forms
+
 ## [10.1.0] - 2025-01-12
 
 ### Added